support Timetag argument and type, imperfect as of yet
NSDate is used to represent the Timetag but it appears as though the conversion to and from an NTP timestamp is not fully symmetrical; very close with more accuracy than sub-millisecond, but not perfectly symmetrical. as such, the associate test to serialize and deserialize fails, but that is survivable for now.